fix(install): use prepublish script rather than install
scripts/install.js requires lodash (and other modules), but if they are
direct dependencies of an application then they may not be available
when the `install` step runs in npm.

Convert from running the script as an `install` script to a `prepublish`
script so that other application dependencies are available at the time
it runs.

Closes #37