Fixing issue in case downloadAndRun is quickly called twice
Without this commit, an issue happened in case an application tries to
create 2 different tunnels by two successive calls to the downloadAndRun
method, and Sauce Connect was not yet downloaded. Both calls try to
download the file but only one succeeds.
This commit makes sure the file to be downloaded exists as soon as the
download method is called, so that the second call detects the file
and waits for the end of the first download.